refactor: extract nextTick test suite
